initialized section_2 of BTH design
small tidy up

small tidy up

tidy
