+++ frontend/package.json
@@ -11,6 +11,7 @@
     "react": "^18.2.0",
     "react-anchor-link-smooth-scroll": "^1.0.12",
     "react-dom": "^18.2.0",
+    "react-hook-form": "^7.42.1",
     "react-icons": "^4.7.1",
     "react-router-dom": "^6.7.0"
   },
