+++ frontend/src/Pages/landing-page/BottomHalfH/Section_2.tsx
@@ -1,0 +1,19 @@
+import HeaderText from '../../../components/HeaderText'
+
+function Section_2() {
+  return (
+    <div className="bg-rose-50">
+      <section id="contactus" className="mx-auto w-5/6 pt-24 pb-222">
+        {/* HEADER */}
+        <HeaderText>Leave us a message</HeaderText>
+        <p className="my-5">
+          Est expedita debitis est reiciendis quia. Et natus molestiae qui
+          temporibus deleniti ut numquam sapiente vel saepe pariatur ab internos
+          error et perspiciatis animi ut minus atque.
+        </p>
+      </section>
+    </div>
+  )
+}
+
+export default Section_2
