+++ frontend/src/Pages/landing-page/index.tsx
@@ -3,6 +3,8 @@ import Navbar from './components/Navbar'
 import HeroSection from './components/HeroSection'
 import FeaturesSection from './components/FeaturesSection'
 import AboutSection from './components/AboutSection'
+import Section_1 from './BottomHalfH/Section_1'
+import Section_2 from './BottomHalfH/Section_2'
 
 function LandingPage() {
   return (
@@ -11,6 +13,8 @@ function LandingPage() {
       <HeroSection />
       <FeaturesSection />
       <AboutSection />
+      <Section_1 />
+      <Section_2 />
     </div>
   )
 }
