+++ frontend/src/Pages/landing-page/types.ts
@@ -1,0 +1,7 @@
+export interface TestimonialType {
+  icon: JSX.Element
+  comment: string
+  name: string
+  location: string
+  stars: number
+}
