+++ frontend/src/index.css
@@ -5,8 +5,6 @@
 @tailwind utilities;
 
 
-
-
 /* buttons syles */
 @layer components {
     .primary1 {
@@ -47,4 +45,4 @@ body,
     width: 100%;
     font-family: "Poppins", sans-serif;
 }
-}
+
