--- frontend/src/shared/types.ts
@@ -1,11 +1,0 @@
-export interface TestimonialType {
-  icon: JSX.Element
-  icon_2: JSX.Element
-  icon_3: JSX.Element
-  icon_4: JSX.Element
-  icon_5: JSX.Element
-  icon_6: JSX.Element
-  comment: string
-  name: string
-  location: string
-}
