+++ frontend/tsconfig.json
@@ -4,7 +4,7 @@
     "module": "ES6",
     "target": "ES6",
     "allowJs": true,
-    "jsx": "react",
+    "jsx": "react-jsx",
     "allowSyntheticDefaultImports": true,
     "moduleResolution": "node"
   },
